refactor: move CLI into app/cli with Justfile
Move cli/ → app/cli/ to match app/ monorepo pattern.
Update Go module path and all internal imports.
Add CLI-specific Justfile with build, test, quality, run recipes.
