+++ app/cli/internal/api/sse.go
