+++ app/cli/internal/api/sse_test.go
