+++ app/cli/internal/command/registry.go
