+++ app/cli/internal/ui/markdown.go
@@ -5,7 +5,7 @@ import (
 
 	"charm.land/lipgloss/v2"
 
-	"github.com/dugshub/stack-bench/cli/internal/ui/theme"
+	"github.com/dugshub/stack-bench/app/cli/internal/ui/theme"
 )
 
 // MarkdownRenderer renders streaming markdown text progressively.
