+++ app/cli/internal/ui/theme/registry.go
